requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
boto3==1.38.1
botocore==1.38.1
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
cnocr==2.3.1
cnstd==1.2.5.2
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.5.0
deprecated==1.2.18
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
doclayout-yolo==0.0.3
durationpy==0.9
effdet==0.4.1
emoji==2.14.1
et-xmlfile==2.0.0
eval-type-backport==0.2.2
evaluate==0.4.3
faiss-cpu==1.10.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2024.12.0
gitdb==4.0.12
gitpython==3.1.44
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-vision==3.10.1
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib-metadata==8.6.1
importlib-resources==6.5.2
jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.24
langchain-aws==0.2.21
langchain-chroma==0.2.3
langchain-community==0.3.22
langchain-core==0.3.55
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.3.33
libmagic==1.0
lightning-utilities==0.14.3
litellm==1.67.2
lxml==5.4.0
markdown==3.8
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
mypy-extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.21.1
openai==1.76.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
optimum==1.24.0
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandoc==2.4
pdf2image==1.17.0
pdfminer==20191125
pdfminer-six==20250416
pi-heif==0.22.0
pikepdf==9.7.0
pillow==11.2.1
pix2text==1.1.3
platformdirs==4.3.7
plumbum==1.9.0
ply==3.11
poppler-utils==0.1.0
posthog==4.0.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyclipper==1.3.0.post6
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.9.1
pygments==2.19.1
pymupdf==1.25.5
pypandoc==1.15
pyparsing==3.2.3
pypdf==5.4.0
pypdfium2==4.30.1
pypika==0.48.9
pyproject-hooks==1.2.0
pyspellchecker==0.8.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytorch-lightning==2.5.1.post0
pytz==2025.2
pyyaml==6.0.2
rapidfuzz==3.13.0
rapidocr-onnxruntime==1.3.25
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
s3transfer==0.12.0
safetensors==0.5.3
scipy==1.15.2
seaborn==0.13.2
sentry-sdk==2.27.0
setproctitle==1.3.5
setuptools==79.0.1
shapely==2.1.0
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.40
starlette==0.46.2
stringzilla==3.12.5
sympy==1.13.3
tenacity==9.1.2
tesseract==0.1.3
thop==0.1.1.post2209072238
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
torch==2.7.0
torchmetrics==1.7.1
torchvision==0.22.0
tqdm==4.67.1
transformers==4.48.3
triton==3.3.0
typer==0.15.2
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
ultralytics==8.3.23
ultralytics-thop==2.0.14
unidecode==1.4.0
unstructured==0.17.2
unstructured-client==0.34.0
unstructured-inference==0.8.10
unstructured-pytesseract==0.3.15
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
wandb==0.19.10
watchfiles==1.0.5
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
xlrd==2.0.1
xlsxwriter==3.2.3
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
